refactor(chat): extract image attachments + thread management
Phase 2e of the 2026-04-20 audit sweep. chat.js was the worst
size×churn file in the codebase (3410 lines × 112 commits in 90
days) — the two natural extraction islands identified in the audit
were image-attachment handling and conversation-thread management,
neither of which touches the delicate streaming hot path.

## New modules

### js/chat-images.js (~212 lines)

Owns the pending-attachment queue, paste/drop/file-picker handlers,
HD-mode toggle, thumbnail generation for chat-history storage, and
the image-lightbox modal. Exports getPendingAttachments() +
hasPendingAttachments() for chat.js's send flow. Calls
window.updateSendButtonState when the queue changes (one back-ref,
matches the codebase's window.fn() convention).

### js/chat-threads.js (~478 lines)

Owns thread-index CRUD (localStorage layout, legacy-flat-chat
migration), the thread-rail UI, message-content search with
debounced body-scan + 30-match cap, and navigate-to-match
highlighting across threads. Calls back into chat.js via window.*
for renderChatMessages / updateChatHeaderTitle / updatePersonalityBar /
loadChatHistory / saveChatHistory / showDiscussContinuePrompt /
renderSavedSummaries / cleanupDiscussionState / getActivePersonality —
all documented at the top of the module.

## chat.js changes

Drops the two extracted blocks (~620 lines removed); adds imports
for the functions chat.js itself still calls; adds a small
updateSendButtonState helper that the image module back-refs via
window. The streaming pipeline, personality flow, onboarding
wizard, and sendChatMessage orchestration stay unchanged.

Final line counts:

  chat.js         3410 → 2779  (-18%)
  chat-images.js     0 → 212
  chat-threads.js    0 → 478

Total LOC holds ~steady (code moved, not cut). What changes is
reviewability: each module has one concern, the streaming hot path
is no longer interleaved with thread-list CRUD, and the churn
pressure on chat.js should drop sharply because image/thread
features land in their own files from here on.

## Test plan

- [x] Full suite green (all assertions pass after the extraction)
- [x] tests/test-image-utils.js updated: "_pendingAttachments lives
      in chat.js" -> "lives in chat-images.js"; added explicit
      assertions that chat.js imports the queue-access API from
      chat-images.js

Bumps to 1.21.9. No user-visible change.
