+++ js/chat.js
@@ -8,7 +8,14 @@ import { getActiveData, getEffectiveRange, getEffectiveRangeForDate, getLatestVa
 import { encryptedSetItem, encryptedGetItem, getEncryptionEnabled } from './crypto.js';
 import { getProfileLocation, setProfileLocation, getLatitudeFromLocation, getLocationCache, latitudeToBand, detectLatitudeWithAI, getProfiles, renameProfile, setProfileSex, setProfileDob } from './profile.js';
 import { callClaudeAPI, hasAIProvider, isAIPaused, setAIPaused, getAIProvider, getActiveModelId, getActiveModelDisplay, supportsVision, supportsWebSearch, isVeniceE2EEActive } from './api.js';
-import { resizeImage, isValidImageType, formatImageBlock, buildVisionContent } from './image-utils.js';
+import { formatImageBlock, buildVisionContent } from './image-utils.js';
+import { getPendingAttachments, hasPendingAttachments, clearAttachments, initChatImageHandlers, updateAttachButtonVisibility, renderAttachmentPreview } from './chat-images.js';
+import {
+  loadChatThreads, saveChatThreadIndex, ensureActiveThread, createNewThread,
+  switchToThread, deleteThread, renameThread, autoNameThread, pruneOldThreads,
+  renderThreadList, invalidateThreadContentCache, filterThreadList,
+  restoreRailState, getChatThreadKey, getChatThreadsKey,
+} from './chat-threads.js';
 import { onChatSaved } from './sync.js';
 import { buildLabContext, getContextSummary, injectLensChunks } from './lab-context.js';
 import { hasLens, queryLens, updateLensIndicator } from './lens.js';
@@ -70,631 +77,27 @@ function createTypewriter(el, typingEl, container) {
   };
 }
 
-// ═══════════════════════════════════════════════
-// IMAGE ATTACHMENTS
-// ═══════════════════════════════════════════════
-const MAX_ATTACHMENTS = 5;
-const THUMB_SIZE = 80;
-let _pendingAttachments = []; // { base64, mediaType, name, previewUrl }
-let _hdMode = localStorage.getItem('labcharts-hd-images') === 'true';
-
-/** Shrink an image to a tiny thumbnail data URL for chat history storage */
-function makeThumbnail(previewUrl, width, height) {
-  return new Promise(resolve => {
-    const img = new Image();
-    img.onload = () => {
-      const scale = THUMB_SIZE / Math.max(width, height);
-      const canvas = document.createElement('canvas');
-      canvas.width = Math.round(width * scale);
-      canvas.height = Math.round(height * scale);
-      canvas.getContext('2d').drawImage(img, 0, 0, canvas.width, canvas.height);
-      resolve(canvas.toDataURL('image/jpeg', 0.6));
-    };
-    img.onerror = () => resolve(null);
-    img.src = previewUrl;
-  });
-}
-
-export function toggleHDMode() {
-  _hdMode = !_hdMode;
-  localStorage.setItem('labcharts-hd-images', _hdMode);
-  const btn = document.getElementById('chat-hd-btn');
-  if (btn) {
-    btn.classList.toggle('active', _hdMode);
-    btn.title = hdTitle();
-  }
-}
-
-function hdTitle() {
-  return _hdMode ? 'HD quality (2048px) — click for standard' : 'Standard quality (1024px) — click for HD';
-}
-
-export async function addImageAttachment(file) {
-  if (!isValidImageType(file.type)) {
-    showNotification('Unsupported image type. Use JPEG, PNG, GIF, or WebP.', 'error');
-    return;
-  }
-  if (_pendingAttachments.length >= MAX_ATTACHMENTS) {
-    showNotification(`Maximum ${MAX_ATTACHMENTS} images per message`, 'error');
-    return;
-  }
-  try {
-    const maxDim = _hdMode ? 2048 : 1024;
-    const quality = _hdMode ? 0.92 : 0.85;
-    const { base64, mediaType, width, height, origWidth, origHeight, quality_warnings } = await resizeImage(file, maxDim, quality);
-    const previewUrl = `data:${mediaType};base64,${base64}`;
-    const thumbUrl = await makeThumbnail(previewUrl, width, height);
-    _pendingAttachments.push({ base64, mediaType, name: file.name, previewUrl, thumbUrl });
-    renderAttachmentPreview();
-    updateSendButtonState();
-    // Warn about image quality issues
-    const longSide = Math.max(origWidth, origHeight);
-    if (longSide < 512) {
-      showNotification(`Low resolution image (${origWidth}×${origHeight}). AI may struggle with fine details.`, 'info', 5000);
-    } else if (longSide < 1024 && _hdMode) {
-      showNotification(`Image is ${origWidth}×${origHeight} — smaller than HD target. Consider using a higher-res photo.`, 'info', 4000);
-    }
-    if (quality_warnings.length > 0) {
-      showNotification(quality_warnings[0], 'info', 5000);
-    }
-  } catch (e) {
-    showNotification('Failed to process image: ' + e.message, 'error');
-  }
-}
-
-export function removeImageAttachment(index) {
-  _pendingAttachments.splice(index, 1);
-  renderAttachmentPreview();
-  updateSendButtonState();
-}
-
-export function renderAttachmentPreview() {
-  const container = document.getElementById('chat-attach-preview');
-  if (!container) return;
-  if (_pendingAttachments.length === 0) {
-    container.innerHTML = '';
-    container.style.display = 'none';
-    return;
-  }
-  container.style.display = 'flex';
-  container.innerHTML = _pendingAttachments.map((att, i) =>
-    `<div class="chat-attach-thumb" title="${escapeHTML(att.name)}">` +
-    `<img src="${att.previewUrl}" alt="${escapeHTML(att.name)}">` +
-    `<button class="chat-attach-remove" onclick="removeImageAttachment(${i})" aria-label="Remove">&times;</button>` +
-    `</div>`
-  ).join('') +
-  `<span class="chat-attach-count">${_pendingAttachments.length}/${MAX_ATTACHMENTS}</span>`;
-}
-
-export function openImageLightbox(src) {
-  const overlay = document.createElement('div');
-  overlay.className = 'chat-lightbox';
-  const img = document.createElement('img');
-  img.src = src;
-  img.alt = 'Full image';
-  overlay.appendChild(img);
-  overlay.addEventListener('click', () => overlay.remove());
-  const close = (e) => { if (e.key === 'Escape') { overlay.remove(); document.removeEventListener('keydown', close); } };
-  document.addEventListener('keydown', close);
-  document.body.appendChild(overlay);
-}
-
-export function clearAttachments() {
-  _pendingAttachments = [];
-  renderAttachmentPreview();
-}
-
+// Image-attachment flow (paste/drop/picker handlers, HD-mode toggle,
+// pending-queue, thumbnail generation) lives in chat-images.js as of
+// v1.21.9. chat.js imports read-only queue access + clearAttachments
+// at the top of the file; chat-images.js exposes the user-facing
+// functions on window directly. Back-reference from chat-images.js
+// into chat.js uses window.updateSendButtonState below.
 function updateSendButtonState() {
   const input = document.getElementById('chat-input');
   const sendBtn = document.getElementById('chat-send-btn');
   if (!sendBtn) return;
-  const hasContent = (input && input.value.trim()) || _pendingAttachments.length > 0;
+  const hasContent = (input && input.value.trim()) || hasPendingAttachments();
   sendBtn.disabled = !hasContent && !_chatAbortController;
 }
+Object.assign(window, { updateSendButtonState });
 
-export function updateAttachButtonVisibility() {
-  const visible = hasAIProvider() && supportsVision();
-  const btn = document.getElementById('chat-attach-btn');
-  if (btn) btn.style.display = visible ? 'flex' : 'none';
-  const hdBtn = document.getElementById('chat-hd-btn');
-  if (hdBtn) {
-    hdBtn.style.display = visible ? 'flex' : 'none';
-    hdBtn.classList.toggle('active', _hdMode);
-    hdBtn.title = hdTitle();
-  }
-}
-
-export function initChatImageHandlers() {
-  const textarea = document.getElementById('chat-input');
-  const chatMessages = document.getElementById('chat-messages');
-  const fileInput = document.getElementById('chat-image-input');
-
-  // Paste handler
-  if (textarea) {
-    textarea.addEventListener('paste', (e) => {
-      const items = e.clipboardData?.items;
-      if (!items) return;
-      for (const item of items) {
-        if (item.type.startsWith('image/')) {
-          e.preventDefault();
-          const file = item.getAsFile();
-          if (file) addImageAttachment(file);
-        }
-      }
-    });
-  }
-
-  // Drag-drop on chat messages area
-  if (chatMessages) {
-    chatMessages.addEventListener('dragover', (e) => {
-      if (!supportsVision()) return;
-      const hasImage = [...e.dataTransfer.types].includes('Files');
-      if (hasImage) {
-        e.preventDefault();
-        e.stopPropagation();
-        chatMessages.classList.add('chat-drop-active');
-      }
-    });
-    chatMessages.addEventListener('dragleave', (e) => {
-      if (!chatMessages.contains(e.relatedTarget)) {
-        chatMessages.classList.remove('chat-drop-active');
-      }
-    });
-    chatMessages.addEventListener('drop', (e) => {
-      e.preventDefault();
-      e.stopPropagation();
-      chatMessages.classList.remove('chat-drop-active');
-      if (!supportsVision()) return;
-      const files = [...e.dataTransfer.files].filter(f => f.type.startsWith('image/'));
-      for (const file of files) addImageAttachment(file);
-    });
-  }
-
-  // File input change
-  if (fileInput) {
-    fileInput.addEventListener('change', (e) => {
-      for (const file of e.target.files) {
-        addImageAttachment(file);
-      }
-      e.target.value = '';
-    });
-  }
-}
-
-// ═══════════════════════════════════════════════
-// THREAD STORAGE HELPERS
-// ═══════════════════════════════════════════════
-const MAX_THREADS = 50;
-
-export function getChatThreadsKey() {
-  return `labcharts-${state.currentProfile}-chat-threads`;
-}
-
-export function getChatThreadKey(threadId) {
-  return `labcharts-${state.currentProfile}-chat-t_${threadId}`;
-}
-
-function generateThreadId() {
-  return 't_' + Date.now().toString(36);
-}
-
-// ═══════════════════════════════════════════════
-// THREAD INDEX CRUD
-// ═══════════════════════════════════════════════
-export function loadChatThreads() {
-  const raw = localStorage.getItem(getChatThreadsKey());
-  if (raw) {
-    try {
-      state.chatThreads = JSON.parse(raw);
-    } catch { state.chatThreads = []; }
-  } else {
-    // Migration: convert legacy flat chat array to a thread
-    state.chatThreads = [];
-    const legacyKey = `labcharts-${state.currentProfile}-chat`;
-    const legacyRaw = localStorage.getItem(legacyKey);
-    if (legacyRaw) {
-      try {
-        const messages = JSON.parse(legacyRaw);
-        if (Array.isArray(messages) && messages.length > 0) {
-          const threadId = 't_migrated';
-          const now = new Date().toISOString();
-          state.chatThreads = [{
-            id: threadId,
-            name: 'Previous Chat',
-            createdAt: now,
-            updatedAt: now,
-            messageCount: messages.length,
-            personality: state.currentChatPersonality || 'default'
-          }];
-          // Write per-thread messages (plaintext — encryption handled by save)
-          localStorage.setItem(getChatThreadKey(threadId), legacyRaw);
-          saveChatThreadIndex();
-          // Leave legacy key in place for rollback safety
-        }
-      } catch {}
-    }
-  }
-}
-
-export function saveChatThreadIndex() {
-  localStorage.setItem(getChatThreadsKey(), JSON.stringify(state.chatThreads));
-  onChatSaved();
-}
-
-export function ensureActiveThread() {
-  if (state.currentThreadId) {
-    const exists = state.chatThreads.find(t => t.id === state.currentThreadId);
-    if (exists) return;
-  }
-  // Pick most recent thread or create new
-  if (state.chatThreads.length > 0) {
-    const sorted = state.chatThreads.slice().sort((a, b) => b.updatedAt.localeCompare(a.updatedAt));
-    state.currentThreadId = sorted[0].id;
-  } else {
-    createNewThread();
-  }
-}
-
-export function createNewThread() {
-  const id = generateThreadId();
-  const now = new Date().toISOString();
-  const p = getActivePersonality();
-  const thread = {
-    id,
-    name: 'New Conversation',
-    createdAt: now,
-    updatedAt: now,
-    messageCount: 0,
-    personality: state.currentChatPersonality || 'default',
-    personalityName: p.name,
-    personalityIcon: p.icon
-  };
-  state.chatThreads.unshift(thread);
-  pruneOldThreads();
-  saveChatThreadIndex();
-  cleanupDiscussionState();
-  // Reset to default personality for new thread
-  state.currentChatPersonality = 'default';
-  localStorage.setItem(`labcharts-${state.currentProfile}-chatPersonality`, 'default');
-  state.currentThreadId = id;
-  state.chatHistory = [];
-  renderChatMessages();
-  updateChatHeaderTitle();
-  updatePersonalityBar();
-  renderThreadList();
-  // Focus input
-  const input = document.getElementById('chat-input');
-  if (input) input.focus();
-}
-
-export async function switchToThread(threadId) {
-  if (threadId === state.currentThreadId) return;
-  // Save current thread messages
-  await saveChatHistory();
-  cleanupDiscussionState();
-  // Switch
-  state.currentThreadId = threadId;
-  await loadChatHistory();
-  // Update thread personality
-  const thread = state.chatThreads.find(t => t.id === threadId);
-  if (thread && thread.personality) {
-    state.currentChatPersonality = thread.personality;
-    localStorage.setItem(`labcharts-${state.currentProfile}-chatPersonality`, thread.personality);
-    updateChatHeaderTitle();
-    updatePersonalityBar();
-  }
-  // Restore discussion state if this thread had an active discussion
-  if (thread && thread.discussionPersonas) {
-    showDiscussContinuePrompt(thread.discussionPersonas, thread.discussionOriginalPersonality);
-  }
-  renderThreadList();
-}
-
-export function deleteThread(threadId) {
-  showConfirmDialog('Delete this conversation? This cannot be undone.', () => {
-    invalidateThreadContentCache();
-    // Remove from index
-    state.chatThreads = state.chatThreads.filter(t => t.id !== threadId);
-    saveChatThreadIndex();
-    // Remove per-thread messages
-    localStorage.removeItem(getChatThreadKey(threadId));
-    // Remove saved summary
-    if (state.importedData.chatSummaries) {
-      state.importedData.chatSummaries = state.importedData.chatSummaries.filter(s => s.threadId !== threadId);
-      saveImportedData();
-    }
-    renderSavedSummaries();
-    // If we deleted the active thread, switch
-    if (state.currentThreadId === threadId) {
-      if (state.chatThreads.length > 0) {
-        state.currentThreadId = state.chatThreads[0].id;
-        loadChatHistory();
-      } else {
-        createNewThread();
-      }
-    }
-    renderThreadList();
-    showNotification('Conversation deleted', 'info');
-  });
-}
-
-export function renameThread(threadId, newName) {
-  const thread = state.chatThreads.find(t => t.id === threadId);
-  if (thread && newName && newName.trim()) {
-    thread.name = newName.trim().slice(0, 60);
-    saveChatThreadIndex();
-    renderThreadList();
-  }
-}
-
-export async function renameThreadPrompt(threadId) {
-  const thread = state.chatThreads.find(t => t.id === threadId);
-  if (!thread) return;
-  const name = await window.showPromptDialog('Rename conversation:', {
-    defaultValue: thread.name,
-    okLabel: 'Rename',
-  });
-  if (name) renameThread(threadId, name);
-}
-
-export function autoNameThread(threadId, firstMessage) {
-  const thread = state.chatThreads.find(t => t.id === threadId);
-  if (!thread || thread.name !== 'New Conversation') return;
-  // Extract first 40 chars from the message, trimmed at word boundary
-  let excerpt = firstMessage.replace(/\s+/g, ' ').trim();
-  if (excerpt.length > 40) {
-    excerpt = excerpt.slice(0, 40);
-    const lastSpace = excerpt.lastIndexOf(' ');
-    if (lastSpace > 20) excerpt = excerpt.slice(0, lastSpace);
-    excerpt += '\u2026';
-  }
-  thread.name = excerpt;
-  saveChatThreadIndex();
-  renderThreadList();
-}
-
-export function pruneOldThreads() {
-  if (state.chatThreads.length <= MAX_THREADS) return;
-  // Sort by updatedAt desc, remove oldest
-  const sorted = state.chatThreads.slice().sort((a, b) => b.updatedAt.localeCompare(a.updatedAt));
-  const toRemove = sorted.slice(MAX_THREADS);
-  for (const t of toRemove) {
-    localStorage.removeItem(getChatThreadKey(t.id));
-  }
-  state.chatThreads = sorted.slice(0, MAX_THREADS);
-  saveChatThreadIndex();
-  if (toRemove.length > 0) {
-    showNotification(`Pruned ${toRemove.length} old conversation(s)`, 'info');
-  }
-}
-
-// ═══════════════════════════════════════════════
-// THREAD RAIL UI
-// ═══════════════════════════════════════════════
-export function renderThreadList(filter) {
-  const list = document.getElementById('chat-thread-list');
-  if (!list) return;
-  let threads = state.chatThreads.slice().sort((a, b) => b.updatedAt.localeCompare(a.updatedAt));
-  if (filter && filter.trim()) {
-    const q = filter.toLowerCase().trim();
-    threads = threads.filter(t => t.name.toLowerCase().includes(q));
-  }
-  if (threads.length === 0) {
-    list.innerHTML = '<div style="padding:12px 10px;font-size:11px;color:var(--text-muted);text-align:center">' +
-      (filter ? 'No matching conversations' : 'No conversations yet') + '</div>';
-    return;
-  }
-  const personalityMap = {};
-  for (const p of CHAT_PERSONALITIES) personalityMap[p.id] = p.icon;
-
-  list.innerHTML = threads.map(t => {
-    const isActive = t.id === state.currentThreadId;
-    const date = new Date(t.updatedAt);
-    const dateStr = formatThreadDate(date);
-    const icon = t.personalityIcon || personalityMap[t.personality] || personalityMap.default || '';
-    const iconTitle = t.personalityName ? ` title="${escapeHTML(t.personalityName)}"` : '';
-    return `<div class="chat-thread-item${isActive ? ' active' : ''}" onclick="switchToThread('${escapeHTML(t.id)}')" data-thread-id="${escapeHTML(t.id)}">
-      <div class="chat-thread-item-name">${escapeHTML(t.name)}</div>
-      <div class="chat-thread-item-meta">
-        <span${iconTitle}>${icon}</span>
-        <span>${dateStr}</span>
-        <span>${t.messageCount} msg${t.messageCount !== 1 ? 's' : ''}</span>
-      </div>
-      <div class="chat-thread-item-actions">
-        <button class="chat-thread-item-action" onclick="event.stopPropagation();renameThreadPrompt('${escapeHTML(t.id)}')" title="Rename">&#9998;</button>
-        <button class="chat-thread-item-action delete" onclick="event.stopPropagation();deleteThread('${escapeHTML(t.id)}')" title="Delete">&#10005;</button>
-      </div>
-    </div>`;
-  }).join('');
-}
-
-function formatThreadDate(date) {
-  const now = new Date();
-  const diff = now - date;
-  const mins = Math.floor(diff / 60000);
-  if (mins < 1) return 'just now';
-  if (mins < 60) return `${mins}m ago`;
-  const hrs = Math.floor(mins / 60);
-  if (hrs < 24) return `${hrs}h ago`;
-  const days = Math.floor(hrs / 24);
-  if (days < 7) return `${days}d ago`;
-  return date.toLocaleDateString('en-US', { month: 'short', day: 'numeric' });
-}
-
-let _threadSearchTimer = null;
-let _threadContentCache = null; // { profileId, threads: Map<threadId, messages[]> }
-
-async function getThreadMessages(threadId) {
-  // Invalidate cache on profile switch
-  if (!_threadContentCache || _threadContentCache.profileId !== state.currentProfile) {
-    _threadContentCache = { profileId: state.currentProfile, threads: new Map() };
-  }
-  if (_threadContentCache.threads.has(threadId)) return _threadContentCache.threads.get(threadId);
-  try {
-    const key = getChatThreadKey(threadId);
-    const raw = getEncryptionEnabled() ? await encryptedGetItem(key) : localStorage.getItem(key);
-    const messages = raw ? JSON.parse(raw) : [];
-    _threadContentCache.threads.set(threadId, messages);
-    return messages;
-  } catch { return []; }
-}
-
-// Invalidate cache when messages change
-export function invalidateThreadContentCache() { _threadContentCache = null; }
-
-export function filterThreadList(value) {
-  if (!value || !value.trim()) {
-    // Clear highlights when search is cleared
-    document.querySelectorAll('.chat-search-mark').forEach(m => m.replaceWith(m.textContent));
-    document.querySelectorAll('.chat-msg-highlight').forEach(m => m.classList.remove('chat-msg-highlight'));
-    renderThreadList();
-    return;
-  }
-  // Instant: filter thread names
-  renderThreadList(value);
-  // Debounced: search message content
-  clearTimeout(_threadSearchTimer);
-  _threadSearchTimer = setTimeout(() => searchThreadContent(value.trim()), 250);
-}
-
-async function searchThreadContent(query) {
-  const q = query.toLowerCase();
-  const results = [];
-  for (const t of state.chatThreads) {
-    const messages = await getThreadMessages(t.id);
-    for (let i = 0; i < messages.length; i++) {
-      const m = messages[i];
-      if (!m.content) continue;
-      const idx = m.content.toLowerCase().indexOf(q);
-      if (idx === -1) continue;
-      // Extract snippet around match
-      const start = Math.max(0, idx - 40);
-      const end = Math.min(m.content.length, idx + q.length + 40);
-      const pre = (start > 0 ? '\u2026' : '') + m.content.slice(start, idx);
-      const match = m.content.slice(idx, idx + q.length);
-      const post = m.content.slice(idx + q.length, end) + (end < m.content.length ? '\u2026' : '');
-      // Store content prefix for verification on jump
-      results.push({ threadId: t.id, threadName: t.name, msgIndex: i, role: m.role, pre, match, post, contentPrefix: m.content.slice(0, 50) });
-      if (results.length >= 30) break;
-    }
-    if (results.length >= 30) break;
-  }
-  // Re-check input hasn't changed
-  const input = document.getElementById('chat-thread-search');
-  if (!input || input.value.trim().toLowerCase() !== q) return;
-  showSearchResults(q, results);
-}
-
-function showSearchResults(query, results) {
-  const list = document.getElementById('chat-thread-list');
-  if (!list) return;
-  if (results.length === 0) {
-    // Append "no message matches" if thread list already shows no results
-    if (list.textContent.includes('No matching')) {
-      list.innerHTML = '<div style="padding:12px 10px;font-size:11px;color:var(--text-muted);text-align:center">No matches in conversations or messages</div>';
-    }
-    return;
-  }
-  const cap = 30;
-  const truncated = results.length >= cap ? `<div style="padding:6px 10px;font-size:10px;color:var(--text-muted);text-align:center">Showing first ${cap} matches</div>` : '';
-  list.innerHTML = `<div class="chat-search-results-label">Messages</div>` +
-    results.map(r => {
-      const icon = r.role === 'user' ? '\uD83D\uDCDD' : '\uD83E\uDD16';
-      return `<div class="chat-search-result" data-prefix="${escapeHTML(r.contentPrefix)}" onclick="jumpToSearchResult('${escapeHTML(r.threadId)}',${r.msgIndex},this.dataset.prefix)">
-        <div class="chat-search-result-thread">${escapeHTML(r.threadName)}</div>
-        <div class="chat-search-result-snippet">${icon} ${escapeHTML(r.pre)}<mark>${escapeHTML(r.match)}</mark>${escapeHTML(r.post)}</div>
-      </div>`;
-    }).join('') + truncated;
-}
-
-export async function jumpToSearchResult(threadId, msgIndex, contentPrefix) {
-  const input = document.getElementById('chat-thread-search');
-  const query = input?.value?.trim() || '';
-  // Switch to thread (re-renders messages if different thread)
-  if (state.currentThreadId !== threadId) {
-    await window.switchToThread(threadId);
-    // Restore search results after thread switch re-rendered the list
-    if (query) searchThreadContent(query);
-  }
-  // Wait for DOM to settle after potential re-render
-  requestAnimationFrame(() => {
-    let msgEl = document.getElementById('chat-msg-' + msgIndex);
-    // Verify we're highlighting the right message (index may have shifted)
-    if (msgEl && contentPrefix && state.chatHistory[msgIndex]) {
-      const actual = (state.chatHistory[msgIndex].content || '').slice(0, 50);
-      if (actual !== contentPrefix) {
-        // Index shifted — find the right message
-        const correctIdx = state.chatHistory.findIndex(m => m.content && m.content.slice(0, 50) === contentPrefix);
-        if (correctIdx !== -1) msgEl = document.getElementById('chat-msg-' + correctIdx);
-        else msgEl = null;
-      }
-    }
-    if (!msgEl) return;
-    // Remove stale marks from previous search
-    document.querySelectorAll('.chat-search-mark').forEach(m => m.replaceWith(m.textContent));
-    document.querySelectorAll('.chat-msg-highlight').forEach(m => m.classList.remove('chat-msg-highlight'));
-    if (query) highlightInMessage(msgEl, query);
-    const mark = msgEl.querySelector('.chat-search-mark');
-    (mark || msgEl).scrollIntoView({ behavior: 'smooth', block: 'center' });
-    msgEl.classList.add('chat-msg-highlight');
-  });
-}
-
-function highlightInMessage(el, query) {
-  const walker = document.createTreeWalker(el, NodeFilter.SHOW_TEXT);
-  const q = query.toLowerCase();
-  const qLen = query.length;
-  const nodes = [];
-  while (walker.nextNode()) nodes.push(walker.currentNode);
-  for (let i = nodes.length - 1; i >= 0; i--) {
-    const node = nodes[i];
-    const text = node.textContent;
-    const lower = text.toLowerCase();
-    // Find all match positions in this node (reverse order)
-    const positions = [];
-    let pos = 0;
-    while ((pos = lower.indexOf(q, pos)) !== -1) {
-      positions.push(pos);
-      pos += qLen;
-    }
-    if (!positions.length) continue;
-    // Split node at each match, in reverse to keep offsets valid
-    const parent = node.parentNode;
-    let remainder = node;
-    for (let j = positions.length - 1; j >= 0; j--) {
-      const idx = positions[j];
-      const current = remainder.textContent;
-      const before = current.slice(0, idx);
-      const match = current.slice(idx, idx + qLen);
-      const after = current.slice(idx + qLen);
-      const mark = document.createElement('mark');
-      mark.className = 'chat-search-mark';
-      mark.textContent = match;
-      if (after) parent.insertBefore(document.createTextNode(after), remainder.nextSibling);
-      parent.insertBefore(mark, remainder.nextSibling);
-      remainder.textContent = before;
-    }
-  }
-}
-
-export function toggleThreadRail() {
-  const rail = document.getElementById('chat-thread-rail');
-  if (!rail) return;
-  const isOpen = rail.classList.toggle('open');
-  localStorage.setItem(`labcharts-${state.currentProfile}-chatRailOpen`, isOpen ? 'true' : 'false');
-}
-
-function restoreRailState() {
-  const rail = document.getElementById('chat-thread-rail');
-  if (!rail) return;
-  const saved = localStorage.getItem(`labcharts-${state.currentProfile}-chatRailOpen`);
-  if (saved === 'true') {
-    rail.classList.add('open');
-  } else {
-    rail.classList.remove('open');
-  }
-}
-
+// Thread management (storage CRUD, rail UI, content search,
+// navigate-to-match highlighting) lives in chat-threads.js as of
+// v1.21.9. chat.js imports the functions it needs from that module;
+// chat-threads.js exposes its HTML-facing functions on window and
+// calls back into chat.js via window.fn() for the render/load/save
+// helpers (renderChatMessages, updateChatHeaderTitle, etc.).
 
 // ═══════════════════════════════════════════════
 // WEB SEARCH
@@ -2548,11 +1951,11 @@ export async function sendChatMessage() {
   const sendBtn = document.getElementById('chat-send-btn');
   const container = document.getElementById('chat-messages');
   const text = input.value.trim();
-  const hasImages = _pendingAttachments.length > 0;
+  const hasImages = hasPendingAttachments();
   if (!text && !hasImages) return;
 
   // Capture attachments before clearing (they're ephemeral)
-  const attachments = hasImages ? [..._pendingAttachments] : [];
+  const attachments = hasImages ? [...getPendingAttachments()] : [];
 
   // Ensure we have a thread
   if (!state.currentThreadId) {
@@ -3348,36 +2751,15 @@ Object.assign(window, {
   getThreadPersonaCount,
   askAIAboutMarker,
   askAIAboutCorrelations,
-  // Thread functions
-  loadChatThreads,
-  saveChatThreadIndex,
-  ensureActiveThread,
-  createNewThread,
-  switchToThread,
-  deleteThread,
-  renameThread,
-  renameThreadPrompt,
-  autoNameThread,
-  pruneOldThreads,
-  renderThreadList,
-  filterThreadList,
-  jumpToSearchResult,
-  toggleThreadRail,
-  // Action bar functions
+  // Thread functions live in chat-threads.js; it does its own Object.assign(window, ...)
+  // Image attachments live in chat-images.js; same pattern.
+  // Action bar
   buildActionBar,
   regenerateLastMessage,
   copyMessage,
   toggleContextDetails,
   getChatWebSearchEnabled,
   setChatWebSearchEnabled,
-  // Image attachments
-  addImageAttachment,
-  toggleHDMode,
-  openImageLightbox,
-  removeImageAttachment,
-  clearAttachments,
-  updateAttachButtonVisibility,
-  initChatImageHandlers,
   setChatNudge,
   updateChatNudge,
   setChatProfileSex,
